requirements.txt
django==4.1.4
django-crispy-forms==1.14.0
crispy-bootstrap5==0.7
